+++ CHANGELOG.md
@@ -1,3 +1,20 @@
+## [1.12.3](https://github.com/kyndryl-design-system/shidoka-applications/compare/v1.12.2...v1.12.3) (2024-04-03)
+
+
+### Bug Fixes
+
+* **checkbox-group:** add on-search and on-limit-toggle events ([a26b7f3](https://github.com/kyndryl-design-system/shidoka-applications/commit/a26b7f3bd1cb827d9bfceb5bb314bbebea234a5a))
+* **tab:** add aria-disabled attribute ([0b90abc](https://github.com/kyndryl-design-system/shidoka-applications/commit/0b90abc6adbbcb79b8b1e55e6b791541c1974ee5))
+
+## [1.12.2](https://github.com/kyndryl-design-system/shidoka-applications/compare/v1.12.1...v1.12.2) (2024-04-03)
+
+
+### Bug Fixes
+
+* **checkbox-group:** don't select all when zero checkboxes ([ded9caf](https://github.com/kyndryl-design-system/shidoka-applications/commit/ded9caf92483fb8b093700020b417fff8169cfb3))
+* **checkbox:** disable individual checkboxes ([008ab30](https://github.com/kyndryl-design-system/shidoka-applications/commit/008ab3088151bda99cac95142591dad4846d8406))
+* **radio-button:** expose disabled prop ([6c7f356](https://github.com/kyndryl-design-system/shidoka-applications/commit/6c7f35626c62e7b81cb0ae34ae5745053729bb97))
+
 ## [1.12.1](https://github.com/kyndryl-design-system/shidoka-applications/compare/v1.12.0...v1.12.1) (2024-04-01)